Added a workaround for "ResistorGate"
An issue exists where the RS485 transceiver will sometimes prevent communicating with the bootloader over UART.
This is ultimately due to the ~RE and DE pins being left floating, which for *some* boards will cause the ~RE pin to bias high enough to enable the transceiver while the bootloader is active. Since the bootloader and RS485 use the same UART, this causes the RS485 transceiver to interfere, preventing flashing over UART.
A hardware fix is needed to completely address this bug, but this change adds an additional software-based workaround for existing boards.

If the user holds down the forward/reverse buttons for about ~5 seconds, the firmware will configure the ~RE and DE pins appropriately, lock their configuration, and then manually jump into the bootloader. Entering the bootloader this way allows us enough control over the state of the pins to disable the transceiver and prevent it from interfering.

Note that if the programming process is interrupted or fails, the device can not be recovered without a SWD programmer. Without this workaround existing boards that exhibit this issue can only be programmed over SWD anyway, so that takes a little bit of sting out of this caveat.